Make it easier to create custom builds of the client
Add a script `./scripts/build-app` which will create a production build
of the client using a given OAuth client ID which can then be served
from the pre-registered URL associated with that OAuth client. This is
useful for early-stage experiments or third-parties wanting to customize the
client in ways that are not suitable for inclusion in the main codebase.

 - Add a template for a minimal version of "app.html", the sidebar app's
   HTML file which is normally hosted by h. At build time, this template
   is processed and emitted into the "build" directory by the
   "build-html" task

 - Make it possible to serve the client from a URL without having to
   bake the URL into the client at build time by using
   `document.currentScript` to resolve the location of the client's
   assets. This avoids the need to rebuild the client if it is moved to
   a different URL.